Code Adjustments
Changed GameSetup to separate out the Init step. I was worried about the game states overwriting while working on the player setup within the same class so they are now separated.

Added more to the player setup section (now in GameSetup).

Started adding player commands and ways to interact with the game state.