GamestateHelper added
Separated read and write to it's own class. No longer is this handled within the other helper classes.

Discovered a possible but where coder would need to be VERY careful with how they ordered helper commands when read/write was inside. Now that it is separated it should be much easier to avoid accidental overwrites.