Add files via upload
Please note that this file will throw an error if used as is. There are 3 entries ("socha", "art", and "rep") that include hashed out comments about what each line means where I thought it could be unclear. Please review, delete the 8 #'ed lines, and reupload for actual use.

art_pt1 and rep_pt1 are a little concerning for me. I'm not sure if 0 will be interpreted incorrectly but I'm not sure how else to specify "point per # of X thing". Maybe this line needs to be revised, I'll post it on the discord too.